Breaking down data file into separate files
